graphics/requirements.txt
paho-mqtt==1.6.1
pygame==2.1.2
